Fix footer alignment problem - issue #1
This fixes the footer alignment problem with some changes to custom.css.

* Use leading dots in the class names to ensure that the names only
apply to classes.
* Set `margin: auto;` in the `gvfooter` class. I suspect that this was
the change that actually enabled centering of the footer components.

The `footer.html` claims that it is out of date but shows no changes in
a compare editor. It is included just in case to keep the repositories
clean.